pavan-updated1

+++ financial_organization_rating_system/lib/src/main/java/com/cg/financial_organization_rating_system/services/UsersServiceImpl.java
@@ -14,84 +14,91 @@ import com.cg.financial_organization_rating_system.entities.Users;
 import com.cg.financial_organization_rating_system.exceptions.InvalidContactDetailsException;
 import com.cg.financial_organization_rating_system.exceptions.InvalidPasswordException;
 import com.cg.financial_organization_rating_system.exceptions.InvalidUserNameException;
+import com.cg.financial_organization_rating_system.exceptions.OrganizationRepEntityNotFoundException;
+import com.cg.financial_organization_rating_system.exceptions.OrganizationRepNotFoundException;
 import com.cg.financial_organization_rating_system.exceptions.UserAddressNotFoundException;
 import com.cg.financial_organization_rating_system.exceptions.UserNotFoundException;
 import com.cg.financial_organization_rating_system.repository.AddressRepository;
 import com.cg.financial_organization_rating_system.repository.LoginRepository;
 import com.cg.financial_organization_rating_system.repository.OrganizationRepRepository;
 import com.cg.financial_organization_rating_system.repository.UsersRepository;
-import com.cg.financial_organization_rating_system.utils.PasswordManagement;
-
 
 @Service
 public class UsersServiceImpl implements UsersService {
 	@Autowired
 	UsersRepository userrepo;
-	
+
 	@Autowired
 	AddressRepository adrsrepo;
-	
+
 	@Autowired
 	OrganizationRepRepository orgreprepo;
-	
+
 	@Autowired
 	LoginRepository loginrepo;
 
-
 	@Override
 	public int userRegistration(UsersRegistrationDto userdto) {
-		
+        
 		Address address = adrsrepo.getAdrsById(userdto.getPincode());
 		Users usr = new Users();
-		if(userdto.getUserName()==null||userdto.getUserName()=="")
+		if (userdto.getUserName() == null || userdto.getUserName() == "")
 			throw new InvalidUserNameException("User name is invalid");
 		usr.setUserName(userdto.getUserName());
-		//if(userdto.getUserContactDetails()!=10||userdto.getUserContactDetails()==0)
-		//	throw new InvalidContactDetailsException("Entered UserContactDetail is Invalid...");
+
+		if (userdto.getUserContactDetails() == 0)
+			throw new InvalidContactDetailsException("Entered UserContactDetail is Invalid...");
+
 		usr.setUserContactDetails(userdto.getUserContactDetails());
-		if(userdto.getPassword().length()!=6||userdto.getPassword()==null) {
-		throw new InvalidPasswordException("Entered password is invalid please re-enter valid password");}
+		if (userdto.getPassword().length() != 6 || userdto.getPassword() == null) {
+			throw new InvalidPasswordException("Entered password is invalid please re-enter valid password");
+		}
 		usr.setPassword(userdto.getPassword());
-		if(address==null)
-		throw new UserAddressNotFoundException("User Address is not found.....");
+		if (address == null)
+			throw new UserAddressNotFoundException("User Address is not found.....");
 		usr.setAddress(address);
-		userrepo.save(usr);
 		
-		LoginDetails login=new LoginDetails();
+		userrepo.save(usr);
+
+		LoginDetails login = new LoginDetails();
 		login.setPassword(userdto.getPassword());
 		login.setLoginId(usr.getUserId());
 		login.setRole("Users");
 		loginrepo.save(login);
-		
+
 		return usr.getUserId();
 	}
 
 	@Override
 	public List<OrganizationRep> browseByEntity() {
-	return orgreprepo.findAll();
+		List<OrganizationRep>orgreplist = orgreprepo.findAll();
+		if(orgreplist.isEmpty())
+		{
+			throw new OrganizationRepEntityNotFoundException("No Organization to display...");
+		}
+		return orgreprepo.findAll();
 	}
 
 	@Override
 	public List<Users> getUserDetails() {
-	return userrepo.findAll();
+		return userrepo.findAll();
 	}
 
 	@Override
 	public Optional<OrganizationRep> browseByEntityId(int orgId) {
+		OrganizationRep orgrep = orgreprepo.getOrgRepById(orgId);
+		if (orgrep.getOrgId() == 0)
+			throw new OrganizationRepNotFoundException("Entered OrgId is not present ....");
 		return orgreprepo.findById(orgId);
-	
+
 	}
 
 	@Override
 	public void deleteUser(int userId) {
 		Users user = userrepo.getUserById(userId);
-		
+		if (user == null)
+			throw new UserNotFoundException("Invalid User Id....");
 		userrepo.deleteById(userId);
-		}
+	}
 
-	
-
-	
-
-	
 }
